Declustering

Declustered statistics are statistics that make use of geospatial coordinates in an attempt to correct potential sampling bias:

The following statistics have geospatial semantics:

Statistics.mean — Method

mean(data, v)
mean(data, v, s)

Declustered mean of geospatial data. Optionally, specify the variable v and the block side s.

Statistics.var — Method

var(data, v)
var(data, v, s)

Declustered variance of geospatial data. Optionally, specify the variable v and the block side s.

Statistics.quantile — Method

quantile(data, v, p)
quantile(data, v, p, s)

Declustered quantile of geospatial data at probability p. Optionally, specify the variable v and the block side s.

A histogram with geospatial semantics is also available where the heights of the bins are adjusted based on the coordinates of the samples:

GeoStatsBase.EmpiricalHistogram — Type

EmpiricalHistogram(sdata, var, [s]; kwargs...)

Spatial histogram of variable var in spatial data sdata. Optionally, specify the block side s and the keyword arguments kwargs for the fit(Histogram, ...) call.