Declustered statistics

Declustered statistics are statistics that make use of geospatial coordinates to correct potential sampling bias:

We provide the following declustered statistics:

Statistics.mean — Method

mean(geotable, column)
mean(geotable, column, side)

Declustered mean of given column in geotable. Optionally, specify the declustering block side.

Statistics.var — Method

var(data, column)
var(data, column, side)

Declustered variance of given column in geotable. Optionally, specify the declustering block side.

Statistics.quantile — Method

quantile(data, column, prob)
quantile(data, column, prob, side)

Declustered quantile of given column in geotable at given probability. Optionally, specify the declustering block side.

A histogram is also available where the heights of the bins are adjusted based on the coordinates of the samples:

GeoStatsBase.EmpiricalHistogram — Type

EmpiricalHistogram(geotable, column, [side]; kwargs...)

Declustered histogram of given column in geotable. Optionally, specify the declustering block side, and the keyword arguments kwargs for the underlying fit(Histogram, ...) call.